[refac]: moved class SSRQSchema to own file
