+++ test/conftest.py
@@ -15,10 +15,10 @@ from ssrq_cli.xml_utils import ext_etree
 
 from utils.constants import SPECIFIED_ELEMENTS, SRC_DIR, XSLTS
 from utils.main import (
-    SSRQSchema,
     load_config,
     odd_factory,
 )
+from utils.SSRQSchema import SSRQSchema
 from utils.SSRQSchemaType import SSRQSchemaType
 
 ElName = str
