+++ utils/main.py
@@ -1,5 +1,4 @@
 import re
-from dataclasses import dataclass
 from typing import Optional
 
 import tomllib
@@ -7,45 +6,14 @@ from saxonche import PySaxonProcessor, PyXdmNode, PyXslt30Processor, PyXsltExecu
 
 from utils.constants import (
     CUR_DIR,
-    DIST_DIR,
     EXAMPLES_DIR,
     SPECIFIED_ELEMENTS,
     SRC_DIR,
     XSLTS,
 )
 from utils.SSRQConfig import SSRQConfig
+from utils.SSRQSchema import SSRQSchema
 from utils.SSRQSchemaType import SSRQSchemaType
-
-OMIT_VERSION: bool = False
-
-
-@dataclass
-class SSRQSchema:
-    version: str
-    name: str
-    compiled_odd: str
-    rng: str
-
-    def store(self) -> None:
-        """Store the compiled ODD and RNG files in the dist directory – the version number is omitted if OMIT_VERSION is True."""
-        from os import makedirs
-        from os.path import exists
-
-        if not exists(DIST_DIR):
-            makedirs(DIST_DIR)
-
-        with open(
-            f"{DIST_DIR}/{self.name}{f'_{self.version}' if OMIT_VERSION is False else ''}.odd",
-            "w",
-            encoding="utf-8",
-        ) as f:
-            f.write(self.compiled_odd)
-        with open(
-            f"{DIST_DIR}/{self.name}{f'_{self.version}' if OMIT_VERSION is False else ''}.rng",
-            "w",
-            encoding="utf-8",
-        ) as f:
-            f.write(self.rng)
 
 
 def load_config() -> Optional[SSRQConfig]:
